c: use correct type in array generation
Earlier, to generate the array, we're calling generate_type on the type
that came to this function which *IS* an array. That creates an infinite
loop and thus causing a stack overflow.

The fix is simple, use type that the array holds (e.g. Int) and not the
type of array itself (which is Array<Int> in this case).