Print error Strings from main verbatim
When returning Result<(), String> from main(), the error message is
escaped. This results in errors like this:

    $ sb build ...
    Error: "Re-declaration of variable \'v\'"

This probably happens because the error from main() is passed to the
debug formatter, but in this case it's better to use the default one.

The solution is to wrap the real main into a function like run(), and
then printing to stderr followed by exit(1) on an error from it.